pyproject.toml
[project]
name = "pro-nlp-finalproject-nlp-05"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite==0.21.0",
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "attrs==25.4.0",
    "authlib==1.6.5",
    "backports-tarfile==1.2.0",
    "beartype==0.22.5",
    "bs4>=0.0.2",
    "cachetools==6.2.1",
    "certifi==2025.10.5",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "click==8.3.0",
    "cryptography==46.0.3",
    "cyclopts==4.2.1",
    "discord>=2.3.2",
    "diskcache==5.6.3",
    "distro==1.9.0",
    "dnspython==2.8.0",
    "docstring-parser==0.17.0",
    "docutils==0.22.2",
    "email-validator==2.3.0",
    "exceptiongroup==1.3.0",
    "fastapi>=0.128.0",
    "fastmcp==2.13.0.2",
    "google-api-python-client>=2.90.0",
    "google-auth>=2.20.0",
    "google-auth-httplib2>=0.1.0",
    "google-auth-oauthlib>=1.1.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.3",
    "idna==3.11",
    "importlib-metadata==8.7.0",
    "jaraco-classes==3.4.0",
    "jaraco-context==6.0.1",
    "jaraco-functools==4.3.0",
    "jiter==0.11.1",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.25.1",
    "jsonschema-path==0.3.4",
    "jsonschema-specifications==2025.9.1",
    "keyring==25.6.0",
    "langchain==1.0.3",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-core==1.0.2",
    "langchain-mcp-adapters==0.1.12",
    "langchain-naver==0.1.1",
    "langchain-openai==1.0.1",
    "langgraph==1.0.2",
    "langgraph-checkpoint==3.0.0",
    "langgraph-checkpoint-sqlite==3.0.0",
    "langgraph-prebuilt==1.0.2",
    "langgraph-sdk==0.2.9",
    "langsmith==0.4.39",
    "markdown-it-py==4.0.0",
    "mcp==1.20.0",
    "mdurl==0.1.2",
    "more-itertools==10.8.0",
    "openai==2.6.1",
    "openapi-pydantic==0.5.1",
    "orjson==3.11.4",
    "ormsgpack==1.11.0",
    "packaging==25.0",
    "pathable==0.4.4",
    "pathvalidate==3.3.1",
    "platformdirs==4.5.0",
    "py-key-value-aio==0.2.8",
    "py-key-value-shared==0.2.8",
    "pycparser==2.23",
    "pydantic==2.12.3",
    "pydantic-core==2.41.4",
    "pydantic-settings==2.11.0",
    "pygments==2.19.2",
    "pyjwt==2.10.1",
    "pyperclip==1.11.0",
    "pyside6>=6.10.2",
    "python-dotenv==1.2.1",
    "python-multipart==0.0.20",
    "pyyaml==6.0.3",
    "qasync>=0.28.0",
    "referencing==0.36.2",
    "regex==2025.10.23",
    "requests==2.32.5",
    "requests-toolbelt==1.0.0",
    "rich==14.2.0",
    "rich-rst==1.3.2",
    "rpds-py==0.28.0",
    "sniffio==1.3.1",
    "sqlalchemy>=2.0.46",
    "sqlite-vec==0.1.6",
    "sse-starlette==3.0.3",
    "starlette==0.50.0",
    "tenacity==9.1.2",
    "tiktoken==0.12.0",
    "tqdm==4.67.1",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "urllib3==2.5.0",
    "uvicorn==0.38.0",
    "websockets==15.0.1",
    "wikipedia>=1.4.0",
    "xxhash==3.6.0",
    "zipp==3.23.0",
    "zstandard==0.25.0",
]
